Fixed all prices being float, now they are all Double

+++ app/src/main/java/db/ecommerce/model/Delivery.java
@@ -15,7 +15,7 @@ public interface Delivery {
     /**
      * @return the priceDelivery
      */
-    float getPriceDelivery();
+    double getPriceDelivery();
 
     /**
      * @return the date
